Initial commit
This is the initial commit of this role. Everything should be working,
whilst not everything has been tested at every edge-case. It is quite
simple, we only need to clone the PrivateBin repository and place the
configuration file. In the future, some more security measures could
take place. Same goes for an example NGINX-config.

Feel free to contribute!